Code/ShipWorks.Shipping/Carriers/FedEx/Api/Rate/Manipulators/Request/FedExRateClientDetailManipulator.cs
using ShipWorks.Data.Model.EntityInterfaces;
using ShipWorks.Shipping.Carriers.FedEx.Api.Environment;
using ShipWorks.Shipping.Carriers.FedEx.WebServices.Rate;

namespace ShipWorks.Shipping.Carriers.FedEx.Api.Rate.Manipulators.Request
{
    /// <summary>
    /// Add client details to a rate request
    /// </summary>
    public class FedExRateClientDetailManipulator : IFedExRateRequestManipulator
    {
        private readonly IFedExSettingsRepository settings;

        /// <summary>
        /// Constructor
        /// </summary>
        public FedExRateClientDetailManipulator(IFedExSettingsRepository settings)
        {
            this.settings = settings;
        }

        /// <summary>
        /// Should the manipulator be applied
        /// </summary>
        public bool ShouldApply(IShipmentEntity shipment, FedExRateRequestOptions options) => true;

        /// <summary>
        /// Manipulates the specified request.
        /// </summary>
        public RateRequest Manipulate(IShipmentEntity shipment, RateRequest request)
        {
            var account = settings.GetAccountReadOnly(shipment);

            request.ClientDetail = FedExRequestManipulatorUtilities.CreateRateClientDetail(account);

            return request;
        }
    }
}
